@@ -19,8 +19,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "phpbb/phpbb": ">=3.3"
+        "php": ">=7.1",
+        "phpbb/phpbb": ">=3.2,<4.0"
     },
     "require-dev": {
         "phpbb/epv": "dev-master"
@@ -28,7 +28,7 @@
     "extra": {
         "display-name": "New Topics Need Approval",
         "soft-require": {
-            "phpbb/phpbb": ">=3.3"
+            "phpbb/phpbb": ">=3.2,<4.0"
         },
         "version-check": {
             "host": "www.phpbb.com",
